requirements.txt
sentence-transformers
uvicorn
ctransformers
langchain
python-box
streamlit
groq
langchain-groq
python-dotenv





# pypdf
# transformers
# einops
# accelerate
# langchain
# bitsandbytes
# pandas
# sentence_transformers
# llama_index
# llama_index_llms_huggingface
